--- src/completion/provider/configuration.ts
@@ -1,13 +1,0 @@
-export interface Config {
-  autoSlash: boolean;
-  mappings: Mapping[];
-  showHiddenFiles: boolean;
-  withExtension: boolean;
-  absolutePathToWorkspace: boolean;
-  filesExclude: { [key: string]: string };
-}
-
-export interface Mapping {
-  key: string;
-  value: string;
-}
