+++ src/completion/setupCompletionItemProvider.ts
@@ -7,11 +7,12 @@ import { appJsonPattern } from '../manifest/utils/parseExpoJson';
 import { subscribeToTsConfigChanges } from './provider/configuration/getTsconfig';
 import { provideCompletionItems } from './provider/provideCompletionItems';
 
+const triggerCharacters = [path.sep, '.', '"'];
+
 export function setupCompletionItemProvider(context: vscode.ExtensionContext) {
   // Subscribe to the ts config changes
   context.subscriptions.push(...subscribeToTsConfigChanges());
 
-  const triggerCharacters = [path.sep, '.', '"'];
   context.subscriptions.push(
     languages.registerCompletionItemProvider(
       appJsonPattern,
